requirements.txt
aiohttp
pytest
pytest-asyncio
PyYAML
openapi-spec-validator